+++ test/lib/perfect_shape/test_point.rb
@@ -6,6 +6,21 @@ require_relative '../../../lib/perfect-shape'
 describe PerfectShape do
   describe PerfectShape::Point do
     it 'constructs with x, y coordinates' do
+      shape = PerfectShape::Point.new(x: 200, y: 150)
+
+      _(shape.x).must_equal 200
+      _(shape.y).must_equal 150
+      _(shape.min_x).must_equal 200
+      _(shape.min_y).must_equal 150
+      _(shape.max_x).must_equal 200
+      _(shape.max_y).must_equal 150
+      _(shape.width).must_equal 0
+      _(shape.height).must_equal 0
+      _(shape.center_x).must_equal 200
+      _(shape.center_y).must_equal 150
+    end
+
+    it 'constructs with x, y coordinates' do
       shape = PerfectShape::Point.new(200, 150)
 
       _(shape.x).must_equal 200